% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AermodEFsForUnpavedRoads.R
\name{calcEFforAermod}
\alias{calcEFforAermod}
\title{calcEFforAermod}
\usage{
calcEFforAermod(EFx, nV, t = units::set_units(1, h), roadLength, roadWidth)
}
\arguments{
\item{EFx}{Numeric. Output from EFpublicRoadUnpaved() or unpavedRoadDustEmissionFactor()}

\item{nV}{Numeric. Average number of vehicles per time unit t}

\item{t}{Units of time}

\item{roadLength}{Numeric with distance unit. Length of road}

\item{roadWidth}{Numeric with distance unit. Width of road}
}
\description{
Uses function 'EFpublicRoadUnpaved' to calculate an emission
 factor suitable for input to Aermod.
}
\examples{
 (1) Calculate the PM2.5 emission factor for a 210 m stretch of road that is 2 m wide,
 has a silt content of 4.1 \%, a surface moisture content of 1.8 \%, and with a
 traffic load of 50 vehicles per hour travelling at 47 km/h on average.
 EFx <- EFpublicRoadUnpaved(s = 4.1, S = 47, M = 1.8, pm = "2.5")
 xAermod <- calcEFforAermod(units::set_units(EFx, g/km),
                        nV = 50, t = units::set_units(1, h),
                        roadLength = units::set_units(210, m),
                        roadWidth =  units::set_units(2, m))
(1) Calculate the daytime PM10 emission factor for a 100 m stretch of road that
is 2 m wide, has a silt content of 6.0 \%, a surface moisture content of 0.7 \%,
and with a daytime (16 hours) traffic load of 110 vehicles in total, travelling
at 63 km/h on average.
EFx <- EFpublicRoadUnpaved(s = 6.0, S = 63, M = 0.7, pm = "10")
xAermod <- calcEFforAermod(units::set_units(EFx, g/km),
                       nV = 110, t = units::set_units(16, h),
                       roadLength = units::set_units(100, m),
                       roadWidth = units::set_units(2, m))
}
