% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AermodEFsForUnpavedRoads.R
\name{EFpublicRoadUnpaved}
\alias{EFpublicRoadUnpaved}
\title{EFpublicRoadUnpaved}
\usage{
EFpublicRoadUnpaved(s = NULL, S = NULL, M = NULL, pm = c("2.5", "10", "30"))
}
\arguments{
\item{s}{Numeric. Surface silt content (%).}

\item{S}{Numeric. Mean vehicle speed (km/h).}

\item{M}{Numeric. Surface moisture content (%).}

\item{pm}{Character. Maximum particulate matter diameter.}
}
\value{
Numeric. Emission factor in gram per kilometer.
}
\description{
Calculates size-specific particulate emissions (in grams) from an
 unpaved public road, per vehicle per kilometer travelled.
}
