% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVehicleDistanceLookup.R
\name{makeVehicleDistanceLookup}
\alias{makeVehicleDistanceLookup}
\title{makeVehicleDistanceLookup}
\usage{
makeVehicleDistanceLookup()
}
\description{
Create a lookup table for the values of K in EPA 42 Ch 13 Section 13.2.1.3
}
\references{
https://www3.epa.gov/ttn/chief/old/ap42/ch13/s021/final/c13s02-1_2002.pdf
}
