% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVKT.R
\name{makeVKT}
\alias{makeVKT}
\title{makeVKT}
\usage{
makeVKT(n = NULL, d = NULL)
}
\arguments{
\item{n}{Integer. Count per hour.}

\item{d}{Numeric. Distance in kilometer}
}
\description{
Calculate the vehicle kilometer travelled
}
