% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisedVehicleCount.R
\name{normalisedVehicleCount}
\alias{normalisedVehicleCount}
\title{normalisedVehicleCount}
\usage{
normalisedVehicleCount(n, l, w, t)
}
\arguments{
\item{n}{Integer. Count of vehicles per unit t}

\item{l}{Numeric or units. Length in meters OR units with distance unit}

\item{w}{Numeric or units. Width in meters OR units with distance unit}

\item{t}{Numeric or units.}
}
\value{
Vehicles per m2 per second [1/m^2/s]
}
\description{
Take vehicle count per timeunit and length and width of road and return vehicle count per m2 per second
}
