% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pavedRoadDustEmissions.R
\name{pavedRoadDustEmissionFactor}
\alias{pavedRoadDustEmissionFactor}
\title{pavedRoadDustEmissionFactor}
\usage{
pavedRoadDustEmissionFactor(
  massUnit = c("g", "lb")[1],
  distanceUnit = c("km", "ml")[1],
  size = c("PM2.5", "PM10", "PM15", "PM30")[2],
  sL = NULL,
  W = NULL,
  precipitationPeriod = NULL,
  P = NULL,
  N = 365
)
}
\arguments{
\item{massUnit}{Character. One of c("g", "lb"). Default: "g"}

\item{distanceUnit}{Character. One of c("km", "ml")}

\item{size}{Character. One of c("PM2.5", "PM10", "PM15", "PM30"). Default:  "PM10"}

\item{sL}{Numeric. road surface silt loading (grams per square meter) (g/m2)}

\item{W}{Numeric. average weight (tons) of the vehicles traveling the road}

\item{precipitationPeriod}{Character. One of c("day", "hour")}

\item{P}{Numeric. Number of “wet” days or hours (see precipitationPeriod) with at least 0.254 mm (0.01 in) of precipitation during the averaging period.}

\item{N}{Numeric. Number of days or hours (see precipitationPeriod) in the averaging period
e.g., 365 for annual, 91 for seasonal, 30 for monthly days or
8760 for annual, 2124 for seasonal, 720 for monthly hours}
}
\value{
Numeric. Particulate emission factor (having units matching the units of k)
}
\description{
Function to calculate dust emissions from paved roads
}
\references{
https://www3.epa.gov/ttn/chief/old/ap42/ch13/s021/final/c13s02-1_2002.pdf
}
