% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpavedRoadDustEmissions.R
\name{unpavedRoadDustEmissionFactor}
\alias{unpavedRoadDustEmissionFactor}
\title{pavedRoadDustEmissionFactor}
\usage{
unpavedRoadDustEmissionFactor(
  size = c("PM2.5", "PM10", "PM30")[2],
  s = NULL,
  W = NULL,
  S = NULL,
  M = NULL,
  type = c("unpaved-industrial", "unpaved-public")[2],
  outunit = c("g/VKT", "lb/VMT")[1]
)
}
\arguments{
\item{size}{Character. One of c("PM2.5", "PM10", "PM15", "PM30"). Default:  "PM10"}

\item{s}{Numeric. surface material silt content (percent)}

\item{W}{Numeric. mean vehicle weight (tons)}

\item{S}{Numeric.mean vehicle speed}

\item{M}{Numeric. Surface material moisture content (percent)}

\item{type}{Character. What type of unpaved road is it? Choose between  "unpaved-industrial" and "unpaved-public"}

\item{massUnit}{Character. One of c("g", "lb"). Default: "g"}

\item{distanceUnit}{Character. One of c("km", "ml")}
}
\value{
Numeric. E. size-specific emission factor (lb/VMT)
}
\description{
Function to calculate dust emissions from unpaved roads
}
\references{
https://www3.epa.gov/ttn/chief/old/ap42/ch13/s021/final/c13s02-1_2002.pdf
}
