% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFleetCLookup.R
\name{makeFleetCLookup}
\alias{makeFleetCLookup}
\title{makeFleetCLookup}
\usage{
makeFleetCLookup(result = c("mile", "km")[2])
}
\arguments{
\item{result}{Character. Units to get values of Emission Factor for Exhaust, Brake Wear and Tire Wearb
Either "mile" or "km"}
}
\description{
Create a lookup table for the values of C in EPA 42 Ch 13 Section 13.2.2 Table 13.2.2-4
}
\references{
https://www3.epa.gov/ttnchie1/ap42/ch13/final/c13s0202.pdf
}
