% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpavedRoadDustEmissions.R
\name{unpavedRoadDustEmissionFactor}
\alias{unpavedRoadDustEmissionFactor}
\title{pavedRoadDustEmissionFactor}
\usage{
unpavedRoadDustEmissionFactor(
  massUnit = c("g", "lb")[1],
  distanceUnit = c("km", "ml")[1],
  size = c("PM2.5", "PM10", "PM15", "PM30")[2],
  sL = NULL,
  W = NULL,
  precipitationPeriod = NULL,
  P = NULL,
  N = 365
)
}
\arguments{
\item{massUnit}{Character. One of c("g", "lb"). Default: "g"}

\item{distanceUnit}{Character. One of c("km", "ml")}

\item{size}{Character. One of c("PM2.5", "PM10", "PM15", "PM30"). Default:  "PM10"}

\item{W}{Numeric. mean vehicle weight (tons)}

\item{E}{Numeric. size-specific emission factor (lb/VMT) s= surface material silt content (%)}

\item{M}{Numeric. surface material moisture content (%)}

\item{S}{Numeric.mean vehicle speed (mph)}

\item{C}{Numeric. emission factor for 1980s vehicle fleet exhaust, brake wear and tire wear.}
}
\value{
Numeric. Particulate emission factor (having units matching the units of k)
}
\description{
Function to calculate dust emissions from paved roads
}
\references{
https://www3.epa.gov/ttn/chief/old/ap42/ch13/s021/final/c13s02-1_2002.pdf
}
